third_party/blink/web_tests/editing/text-iterator/findString-shadow-roots-expected.txt

Find text in shadow roots.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting testSimpleShadow()...
PASS testRunner.findString("SHADOW", []) is true
PASS testRunner.findString("HADO", []) is true
PASS testRunner.findString("SHA", []) is true
PASS testRunner.findString("DOW", []) is true
PASS testRunner.findString("DOCUMENT", []) is false
PASS testRunner.findString("DOC", []) is false
PASS testRunner.findString("CUM", []) is false
PASS testRunner.findString("ENT", []) is false
PASS testRunner.findString("BEFORE", []) is true
PASS testRunner.findString("AFTER", []) is true
PASS testRunner.findString("BEFOREDOC", []) is false
PASS testRunner.findString("ENTAFTER", []) is false
PASS testRunner.findString("ORESHADOW", []) is false
PASS testRunner.findString("SHADOWAFTER", []) is false
Finished testSimpleShadow().

Starting testNestedShadows()...
PASS testRunner.findString("DOCUMENT", []) is false
PASS testRunner.findString("OUTER", []) is true
PASS testRunner.findString("SHADOW", []) is false
PASS testRunner.findString("INNER", []) is true
Finished testNestedShadows().

Starting testDistribution()...
PASS testRunner.findString("BEFORE", []) is true
PASS testRunner.findString("DOCUMENT", []) is true
PASS testRunner.findString("AFTER", []) is true
PASS testRunner.findString("BEFOREDOCUMENTAFTER", []) is true
PASS testRunner.findString("DOCUMENTBEFORE", []) is false
PASS testRunner.findString("AFTERDOCUMENT", []) is false
Finished testDistribution().

PASS successfullyParsed is true

TEST COMPLETE
